+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.table.lastInsertId.R
@@ -1,10 +1,11 @@
 concerto.table.lastInsertId <-
-function(){
+function(connection = NULL){
   id = NULL
+  if(is.null(connection)) { connection = concerto$connection }
   if(concerto$connectionParams$driver == "pdo_sqlsrv") {
     id = concerto$sqlsrv_last_insert_id
   } else {
-    id = dbGetQuery(concerto$connection, "SELECT LAST_INSERT_ID();")[1,1]
+    id = dbGetQuery(connection, "SELECT LAST_INSERT_ID();")[1,1]
   }
   return(id)
 }