+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.table.query.R
@@ -1,5 +1,6 @@
 concerto.table.query <-
-function(sql, params=list(), n=-1){
+function(sql, params=list(), n=-1, connection = NULL){
+  if(is.null(connection)) { connection = concerto$connection }
   sql <- gsub("^\\s+|\\s+$", "", sql)
   sql <- concerto.table.insertParams(sql, params)
 
@@ -8,19 +9,19 @@ function(sql, params=list(), n=-1){
   result <- NULL
   output <- NULL
   if(startsWith(toupper(sql), "SELECT") || startsWith(toupper(sql), "SHOW")) {
-    result <- dbSendQuery(concerto$connection, sql)
+    result <- dbSendQuery(connection, sql)
     output <- fetch(result, n=n)
   } else if(startsWith(toupper(sql), "INSERT")) {
     if(concerto$connectionParams$driver == "pdo_sqlsrv") {
-         result <- dbSendQuery(concerto$connection, paste0(sql,"; SELECT SCOPE_IDENTITY();"))
+         result <- dbSendQuery(connection, paste0(sql,"; SELECT SCOPE_IDENTITY();"))
          output <- fetch(result, n=1)[1,1]
          concerto$sqlsrv_last_insert_id <<- output
     } else {
-        result <- dbSendStatement(concerto$connection, sql)
+        result <- dbSendStatement(connection, sql)
         output <- dbGetRowsAffected(result)
     }
   } else {
-    result <- dbSendStatement(concerto$connection, sql)
+    result <- dbSendStatement(connection, sql)
     output <- dbGetRowsAffected(result)
   }